requirements.txt
Flask
flask-cors
mysql-connector-python
bcrypt
pyotp
python-dotenv
